+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2017-11-07
+### Added
+- Add new functions to create a file and and directory
+- Add new comments to save and load folder function
+- Add new functions to create message boxes for error, warning, information and questions
+
 ## 0.1.0 - 2017-06-02
 ### Added
 - Init commit and push of different elements to new repository
 
-[Unreleased]: https://gitlab.ecklebe.de/open-source/pylibcklb/compare/v0.1.0...master
+[Unreleased]: https://gitlab.ecklebe.de/open-source/pylibcklb/compare/v0.2.0...master
+[0.2.0]: https://gitlab.ecklebe.de/open-source/pylibcklb/compare/v0.1.0...v0.2.0